refactor: OpenSearch 저장소를 service.opensearch 패키지로 이동
